# src/openapi.yaml
openapi: '3.1.0'
info:
  title: WIZBI Control Plane API
  version: '1.0.0'
  description: |
    REST API for the WIZBI Control Plane — manages organizations, projects, 
    templates, users, and GCP/GitHub infrastructure provisioning.
    
    **Authentication**: All endpoints require either:
    - `X-Firebase-ID-Token` header (for web UI)
    - `X-API-Key` header (for programmatic/agent access)
    
    **MCP Server**: Available at `/api/mcp/sse` for AI agent integration.
  license:
    name: MIT
servers:
  - url: /api
    description: Local / deployed server
components:
  securitySchemes:
    firebaseToken:
      type: apiKey
      in: header
      name: X-Firebase-ID-Token
      description: Firebase ID Token for web UI auth
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
...
security:
  - firebaseToken: []
  - apiKey: []
paths:
...
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project details
    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Deletion started
...
  /github/templates:
    get:
      tags: [Templates]
      summary: List available templates
      responses:
        '200':
          description: List of template repositories
    post:
      tags: [Templates]
      summary: Create new template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Template created
  /github/templates/{repoName}:
    put:
      tags: [Templates]
      summary: Update template description
      parameters:
        - name: repoName
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
      responses:
        '200':
          description: Template updated
    delete:
      tags: [Templates]
      summary: Delete template
      parameters:
        - name: repoName
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Template deleted

  # ─── Users ─────────────────────────────────────────
...
  /api-keys:
    get:
      tags: [API Keys]
      summary: List API keys (masked)
      responses:
        '200':
          description: List of API keys
    post:
      tags: [API Keys]
      summary: Create new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                scopes: { type: array, items: { type: string } }
      responses:
        '201':
          description: API key created (raw key returned only here)
  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke API key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: API key revoked

  # ─── GitHub Setup ──────────────────────────────────
...